Add extra unit tests for product

+++ backend/ApiDesignPatterns.Api/Product/Tests/TestHelpers/Builders/ProductTestDataBuilder.cs
@@ -50,7 +50,7 @@ public class ProductTestDataBuilder
         _safetyWarnings = fixture.Create<string>();
     }
 
-    public ProductTestDataBuilder WithId(int id)
+    public ProductTestDataBuilder WithId(long id)
     {
         _id = id;
         return this;
